influxdb metrics review
cherry picked 6ed626ee7160c04c2d409f3ba5dc6ba555524578

+++ basenode/src/main/java/io/coti/basenode/services/interfaces/IConfirmationService.java
@@ -28,9 +28,9 @@ public interface IConfirmationService {
 
     void shutdown();
 
-    long getWaitingDspConsensusResultsMapSize();
+    int getWaitingDspConsensusResultsMapSize();
 
-    long getWaitingMissingTransactionIndexesSize();
+    int getWaitingMissingTransactionIndexesSize();
 
     int getQueueSize();
 }
